app/pyproject.toml
[tool.black]
target-version = ["py312"]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
azure-identity = "^1.15.0"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.14"
